split schema representation internals
