--- .changeset/rename-schema-representation-artifact-code.md
@@ -1,5 +1,0 @@
----
-"effect": patch
----
-
-Rename `SchemaRepresentation.Artifact.generation` to `code` for consistency with the generated value it contains.
