+++ .changeset/schema-representation-refactoring.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+Rename `SchemaRepresentation.Artifact.generation` to `code` and the representation conversion APIs to `toRepresentation`, `toRepresentations`, `fromRepresentation`, and `fromRepresentations`.
