+++ packages/effect/src/internal/schema/toRepresentation.ts
@@ -1,0 +1,535 @@
+import * as Arr from "../../Array.ts"
+import * as Equal from "../../Equal.ts"
+import type * as Schema from "../../Schema.ts"
+import * as SchemaAST from "../../SchemaAST.ts"
+import type * as SchemaRepresentation from "../../SchemaRepresentation.ts"
+import * as InternalRecord from "../record.ts"
+import * as InternalAnnotations from "./annotations.ts"
+import * as InternalSchema from "./schema.ts"
+
+/** @internal */
+export function toRepresentation(
+  ast: SchemaAST.AST,
+  encoded = false
+): SchemaRepresentation.Document {
+  const { references, representations } = toRepresentations([ast], encoded)
+  return { representation: representations[0], references }
+}
+
+/** @internal */
+export function toRepresentations(
+  asts: readonly [SchemaAST.AST, ...Array<SchemaAST.AST>],
+  encoded = false
+): SchemaRepresentation.MultiDocument {
+  return encoded
+    ? lowerASTs(asts, [], true)
+    : lowerASTs(Arr.map(asts, (ast) => SchemaAST.toType(ast)), [])
+}
+
+type RepresentationAnnotation = SchemaRepresentation.RepresentationAnnotation<SchemaRepresentation.Representation>
+
+function annotationsField<A>(annotations: A | undefined): { readonly annotations: A } | undefined {
+  return annotations === undefined ? undefined : { annotations }
+}
+
+function projectRepresentationAnnotation(
+  input: RepresentationAnnotation
+): RepresentationAnnotation {
+  return input.schemas === undefined
+    ? input
+    : { ...input, schemas: input.schemas.map(projectRepresentation) }
+}
+
+function projectAnnotationBag(
+  input: Readonly<Record<string, unknown>> | undefined,
+  stripStringContent = false
+): Schema.Annotations.Annotations | undefined {
+  if (input === undefined) return undefined
+
+  const out: Record<string, unknown> = {}
+  for (const [key, value] of Object.entries(input)) {
+    if (
+      value === undefined ||
+      (stripStringContent && (key === "contentMediaType" || key === "contentSchema"))
+    ) {
+      continue
+    }
+    if (SchemaAST.isJson(value)) InternalRecord.set(out, key, value)
+  }
+
+  return Object.keys(out).length === 0
+    ? undefined
+    : out
+}
+
+function projectCheck(check: SchemaRepresentation.Check): SchemaRepresentation.Check {
+  switch (check._tag) {
+    case "Filter": {
+      const representation = check.representation === undefined
+        ? undefined
+        : projectRepresentationAnnotation(check.representation)
+      return {
+        _tag: "Filter",
+        aborted: check.aborted,
+        ...(representation === undefined ? undefined : { representation }),
+        ...annotationsField(projectAnnotationBag(check.annotations))
+      }
+    }
+    case "FilterGroup": {
+      const representation = check.representation === undefined
+        ? undefined
+        : projectRepresentationAnnotation(check.representation)
+      return {
+        _tag: "FilterGroup",
+        checks: Arr.map(check.checks, projectCheck),
+        ...(representation === undefined ? undefined : { representation }),
+        ...annotationsField(projectAnnotationBag(check.annotations))
+      }
+    }
+  }
+}
+
+function projectRepresentation(
+  representation: SchemaRepresentation.Representation
+): SchemaRepresentation.Representation {
+  if (representation._tag === "Reference") return representation
+
+  const annotations = projectAnnotationBag(representation.annotations, representation._tag === "String")
+  const checks = representation.checks.map(projectCheck)
+
+  switch (representation._tag) {
+    case "Declaration":
+      return {
+        _tag: "Declaration",
+        ...(representation.representation === undefined
+          ? undefined
+          : { representation: projectRepresentationAnnotation(representation.representation) }),
+        typeParameters: representation.typeParameters.map(projectRepresentation),
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "Suspend":
+      return {
+        _tag: "Suspend",
+        checks: [],
+        thunk: projectRepresentation(representation.thunk),
+        ...annotationsField(annotations)
+      }
+    case "Null":
+    case "Undefined":
+    case "Void":
+    case "Never":
+    case "Unknown":
+    case "Any":
+    case "Number":
+    case "Boolean":
+    case "BigInt":
+    case "Symbol":
+    case "ObjectKeyword":
+      return {
+        _tag: representation._tag,
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "Literal":
+      return {
+        _tag: "Literal",
+        literal: representation.literal,
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "UniqueSymbol":
+      return {
+        _tag: "UniqueSymbol",
+        symbol: representation.symbol,
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "Enum":
+      return {
+        _tag: "Enum",
+        enums: representation.enums,
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "String":
+      return {
+        _tag: "String",
+        checks,
+        ...(representation.contentMediaType === undefined
+          ? undefined
+          : { contentMediaType: representation.contentMediaType }),
+        ...(representation.contentSchema === undefined
+          ? undefined
+          : { contentSchema: projectRepresentation(representation.contentSchema) }),
+        ...annotationsField(annotations)
+      }
+    case "TemplateLiteral":
+      return {
+        _tag: "TemplateLiteral",
+        parts: representation.parts.map(projectRepresentation),
+        checks,
+        ...annotationsField(annotations)
+      }
+    case "Arrays": {
+      return {
+        _tag: "Arrays",
+        elements: representation.elements.map((element) => ({
+          isOptional: element.isOptional,
+          type: projectRepresentation(element.type),
+          ...annotationsField(projectAnnotationBag(element.annotations))
+        })),
+        rest: representation.rest.map(projectRepresentation),
+        checks,
+        ...annotationsField(annotations)
+      }
+    }
+    case "Objects": {
+      const propertySignatures = representation.propertySignatures.map(
+        (property): SchemaRepresentation.PropertySignature => ({
+          name: property.name,
+          isOptional: property.isOptional,
+          isMutable: property.isMutable,
+          type: projectRepresentation(property.type),
+          ...annotationsField(projectAnnotationBag(property.annotations))
+        })
+      )
+      const indexSignatures = representation.indexSignatures.map(
+        (index): SchemaRepresentation.IndexSignature => ({
+          parameter: projectRepresentation(index.parameter),
+          type: projectRepresentation(index.type)
+        })
+      )
+      return {
+        _tag: "Objects",
+        propertySignatures,
+        indexSignatures,
+        checks,
+        ...annotationsField(annotations)
+      }
+    }
+    case "Union":
+      return {
+        _tag: "Union",
+        types: representation.types.map(projectRepresentation),
+        mode: representation.mode,
+        checks,
+        ...annotationsField(annotations)
+      }
+  }
+}
+
+function projectReferences(references: SchemaRepresentation.References): SchemaRepresentation.References {
+  const out: Record<string, SchemaRepresentation.Representation> = {}
+  for (const [key, value] of Object.entries(references)) {
+    const representation = projectRepresentation(value)
+    InternalRecord.set(out, key, representation)
+  }
+  return out
+}
+
+/** @internal */
+export function projectDocument(
+  document: SchemaRepresentation.Document
+): SchemaRepresentation.Document {
+  const representation = projectRepresentation(document.representation)
+  const references = projectReferences(document.references)
+  return { representation, references }
+}
+
+/** @internal */
+export function projectMultiDocument(
+  document: SchemaRepresentation.MultiDocument
+): SchemaRepresentation.MultiDocument {
+  const representations = Arr.map(document.representations, projectRepresentation)
+  const references = projectReferences(document.references)
+  return { representations, references }
+}
+
+/** @internal */
+export function fromSchemaMultiDocument(
+  document: SchemaRepresentation.SchemaMultiDocument
+): SchemaRepresentation.MultiDocument {
+  const definitions = Object.entries(document.definitions).map(([key, schema]) => {
+    const original = schema.ast
+    const encoded = SchemaAST.toEncoded(original)
+    const body = SchemaAST.isSuspend(encoded) ? encoded.thunk() : encoded
+    return { key, original, encoded, body }
+  })
+  const asts = Arr.map(document.schemas, (schema) => SchemaAST.toEncoded(schema.ast))
+  return lowerASTs(asts, definitions)
+}
+
+interface ExternalDefinition {
+  readonly key: string
+  readonly original: SchemaAST.AST
+  readonly encoded: SchemaAST.AST
+  readonly body: SchemaAST.AST
+}
+
+function lowerASTs(
+  asts: readonly [SchemaAST.AST, ...Array<SchemaAST.AST>],
+  externalDefinitions: ReadonlyArray<ExternalDefinition>,
+  encoded = false
+): SchemaRepresentation.MultiDocument {
+  const references: Record<string, SchemaRepresentation.Representation> = {}
+  const referenceMap = new Map<SchemaAST.AST, string>()
+  const uniqueReferences = new Set(externalDefinitions.map((definition) => definition.key))
+  const visiting = new Set<SchemaAST.AST>()
+  const externalBodyReferences = new Map<SchemaAST.AST, string | null>()
+
+  for (const definition of externalDefinitions) {
+    referenceMap.set(definition.original, definition.key)
+    referenceMap.set(definition.encoded, definition.key)
+    externalBodyReferences.set(
+      definition.body,
+      externalBodyReferences.has(definition.body) ? null : definition.key
+    )
+  }
+
+  const representations = Arr.map(asts, (ast) => recur(ast))
+
+  for (const definition of externalDefinitions) {
+    references[definition.key] = recur(definition.body, definition.key)
+  }
+
+  return { representations, references }
+
+  function generateReference(prefix: string): string {
+    let candidate = prefix
+    let suffix = 0
+    while (uniqueReferences.has(candidate)) {
+      candidate = `${prefix}${++suffix}`
+    }
+    uniqueReferences.add(candidate)
+    return candidate
+  }
+
+  function recur(
+    ast: SchemaAST.AST,
+    ownedReference?: string,
+    inheritedIdentifier?: string
+  ): SchemaRepresentation.Representation {
+    let found = referenceMap.get(ast)
+    if (found === undefined && SchemaAST.isSuspend(ast)) {
+      const bodyReference = externalBodyReferences.get(ast.thunk())
+      if (bodyReference !== undefined && bodyReference !== null) {
+        found = bodyReference
+        referenceMap.set(ast, bodyReference)
+      }
+    }
+    if (found !== undefined && found !== ownedReference) {
+      return { _tag: "Reference", $ref: found }
+    }
+
+    let identifier = inheritedIdentifier
+    if (encoded) {
+      const projected = SchemaAST.getLastEncoding(ast)
+      identifier = InternalAnnotations.resolveIdentifier(ast) ?? identifier
+      if (projected !== ast) {
+        return recur(projected, ownedReference, identifier)
+      }
+    }
+
+    identifier = ownedReference === undefined ? identifier ?? InternalAnnotations.resolveIdentifier(ast) : undefined
+    if (identifier !== undefined) {
+      const reference = generateReference(identifier)
+      referenceMap.set(ast, reference)
+      const representation = on(ast)
+      const existing = references[identifier]
+      if (existing !== undefined && Equal.equals(representation, existing)) {
+        referenceMap.set(ast, identifier)
+        return { _tag: "Reference", $ref: identifier }
+      }
+      references[reference] = representation
+      return { _tag: "Reference", $ref: reference }
+    }
+
+    if (visiting.has(ast)) {
+      const reference = generateReference(`${ast._tag}_`)
+      referenceMap.set(ast, reference)
+      return { _tag: "Reference", $ref: reference }
+    }
+
+    visiting.add(ast)
+    const representation = on(ast)
+    visiting.delete(ast)
+
+    const reference = referenceMap.get(ast)
+    if (reference !== undefined && reference !== ownedReference) {
+      references[reference] = representation
+      return { _tag: "Reference", $ref: reference }
+    }
+
+    return representation
+  }
+
+  function on(ast: SchemaAST.AST): SchemaRepresentation.Representation {
+    switch (ast._tag) {
+      case "Declaration":
+        return {
+          _tag: "Declaration",
+          typeParameters: ast.typeParameters.map((ast) => recur(ast)),
+          checks: fromChecks(ast.checks),
+          ...fromRepresentationAnnotations(ast.annotations)
+        }
+      case "Null":
+      case "Undefined":
+      case "Void":
+      case "Never":
+      case "Unknown":
+      case "Any":
+      case "Boolean":
+      case "Symbol":
+      case "ObjectKeyword":
+        return {
+          _tag: ast._tag,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "String": {
+        const { contentMediaType, contentSchema, ...annotations } = ast.annotations ?? {}
+        return {
+          _tag: "String",
+          checks: fromChecks(ast.checks),
+          ...(Object.keys(annotations).length === 0 ? undefined : { annotations }),
+          ...(typeof contentMediaType === "string" ? { contentMediaType } : undefined),
+          ...(SchemaAST.isAST(contentSchema) ? { contentSchema: recur(SchemaAST.toType(contentSchema)) } : undefined)
+        }
+      }
+      case "Number":
+      case "BigInt":
+        return {
+          _tag: ast._tag,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Literal":
+        return {
+          _tag: "Literal",
+          literal: ast.literal,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "UniqueSymbol":
+        return {
+          _tag: "UniqueSymbol",
+          symbol: ast.symbol,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Enum":
+        return {
+          _tag: "Enum",
+          enums: ast.enums,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "TemplateLiteral":
+        return {
+          _tag: "TemplateLiteral",
+          parts: ast.parts.map((ast) => recur(ast)),
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Arrays":
+        return {
+          _tag: "Arrays",
+          elements: ast.elements.map((element) => {
+            const projected = encoded ? SchemaAST.getLastEncoding(element) : element
+            const annotations = projected.context?.annotations
+            return {
+              isOptional: SchemaAST.isOptional(projected),
+              type: recur(element),
+              ...(annotations === undefined ? undefined : { annotations })
+            }
+          }),
+          rest: ast.rest.map((ast) => recur(ast)),
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Objects":
+        return {
+          _tag: "Objects",
+          propertySignatures: ast.propertySignatures.map((property) => {
+            const projected = encoded ? SchemaAST.getLastEncoding(property.type) : property.type
+            const annotations = projected.context?.annotations
+            return {
+              name: property.name,
+              type: recur(property.type),
+              isOptional: SchemaAST.isOptional(projected),
+              isMutable: SchemaAST.isMutable(projected),
+              ...(annotations === undefined ? undefined : { annotations })
+            }
+          }),
+          indexSignatures: ast.indexSignatures.map((index) => ({
+            parameter: recur(index.parameter),
+            type: recur(index.type)
+          })),
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Union":
+        return {
+          _tag: "Union",
+          types: InternalSchema.jsonReorder(ast.types).map((ast) => recur(ast)),
+          mode: ast.mode,
+          checks: fromChecks(ast.checks),
+          ...annotationsField(ast.annotations)
+        }
+      case "Suspend":
+        return {
+          _tag: "Suspend",
+          checks: [],
+          thunk: recur(ast.thunk()),
+          ...annotationsField(ast.annotations)
+        }
+    }
+  }
+
+  function fromChecks(
+    checks: readonly [SchemaAST.Check<any>, ...Array<SchemaAST.Check<any>>] | undefined
+  ): Array<SchemaRepresentation.Check> {
+    return checks?.map(fromCheck) ?? []
+  }
+
+  function fromCheck(
+    check: SchemaAST.Check<any>
+  ): SchemaRepresentation.Check {
+    switch (check._tag) {
+      case "Filter":
+        return {
+          _tag: "Filter",
+          aborted: check.aborted,
+          ...fromRepresentationAnnotations(check.annotations)
+        }
+      case "FilterGroup":
+        return {
+          _tag: "FilterGroup",
+          checks: Arr.map(check.checks, fromCheck),
+          ...fromRepresentationAnnotations(check.annotations)
+        }
+    }
+  }
+
+  function fromRepresentationAnnotations<
+    A extends Schema.Annotations.Annotations & {
+      readonly representation?: SchemaRepresentation.RepresentationAnnotation<SchemaAST.AST> | undefined
+    }
+  >(annotations: A | undefined): {
+    readonly representation?: RepresentationAnnotation | undefined
+    readonly annotations?: Omit<A, "representation"> | undefined
+  } {
+    if (annotations === undefined) return {}
+    const { representation, ...ordinary } = annotations
+    const projected = representation === undefined
+      ? undefined
+      : representation.schemas === undefined
+      ? representation as RepresentationAnnotation
+      : { ...representation, schemas: representation.schemas.map((schema) => recur(SchemaAST.toType(schema))) }
+    return {
+      ...(projected === undefined ? undefined : { representation: projected }),
+      ...(Object.keys(ordinary).length === 0 ? undefined : { annotations: ordinary })
+    }
+  }
+}
